Delete test6
Testing that GitHub push still works after seeing this message in
Configure >> Build Triggers >> Build when a change is pushed to GitHub :

"Hook for repo BitMaven/tuts on github.com failed to be registered or were removed. More info can be found on global manage page. This message will be dismissed if Jenkins receives a PING event from repo or repo will be ignored in global configuration."